add tcc demo

+++ at/src/main/java/com/qxdn/dtdemo/service/AccountService.java
@@ -2,5 +2,5 @@ package com.qxdn.dtdemo.service;
 
 public interface AccountService {
 
-    void debit(Long userId,Integer money);
+    void debit(Long userId,Long money);
 }
